Remove redundant duplicate test file, generify test to work better with time zone displays
